+++ frontend/src/components/Sidebar/Sidebar.types.ts
@@ -1,0 +1,8 @@
+type Node = {
+  text: string;
+  level: number;
+  open?: boolean;
+  children?: Node[];
+};
+
+export type { Node };
