--- frontend/src/components/TextEditor.tsx
@@ -1,155 +1,0 @@
-/* eslint-disable react/button-has-type */
-// Import React dependencies.
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-// Import the Slate editor factory.
-import { createEditor, Node, Transforms, Editor, Text } from 'slate';
-
-// Import the Slate components and React plugin.
-import {
-  Slate,
-  Editable,
-  withReact,
-  RenderLeafProps,
-  RenderElementProps,
-} from 'slate-react';
-
-// Define our own custom set of helpers.
-const CustomEditor = {
-  isBoldMarkActive(editor: Editor) {
-    const [match] = Editor.nodes(editor, {
-      match: (n) => n.bold === true,
-      universal: true,
-    });
-
-    return !!match;
-  },
-
-  isCodeBlockActive(editor: Editor) {
-    const [match] = Editor.nodes(editor, {
-      match: (n) => n.type === 'code',
-    });
-
-    return !!match;
-  },
-
-  toggleBoldMark(editor: Editor) {
-    const isActive = CustomEditor.isBoldMarkActive(editor);
-    Transforms.setNodes(
-      editor,
-      { bold: isActive ? null : true },
-      { match: (n) => Text.isText(n), split: true }
-    );
-  },
-
-  toggleCodeBlock(editor: Editor) {
-    const isActive = CustomEditor.isCodeBlockActive(editor);
-    Transforms.setNodes(
-      editor,
-      { type: isActive ? null : 'code' },
-      { match: (n) => Editor.isBlock(editor, n) }
-    );
-  },
-};
-
-const Leaf = (props: RenderLeafProps) => (
-  <span
-    {...props.attributes}
-    style={{ fontWeight: props.leaf.bold ? 'bold' : 'normal' }}
-  >
-    {props.children}
-  </span>
-);
-
-const CodeElement = (props: RenderElementProps) => (
-  <pre {...props.attributes}>
-    <code>{props.children}</code>
-  </pre>
-);
-
-const DefaultElement = (props: RenderElementProps) => (
-  <p {...props.attributes}>{props.children}</p>
-);
-
-const TextEditor = () => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const editor = useMemo(() => withReact(createEditor()), []);
-  // Add the initial value when setting up our state. (taken from local storage commented)
-  const [value, setValue] = useState<Node[]>(
-    // JSON.parse(localStorage.getItem('content')!) ||
-    [
-      {
-        type: 'paragraph',
-        children: [{ text: 'A line of text in a paragraph.' }],
-      },
-    ]
-  );
-
-  const renderElement = useCallback((props) => {
-    switch (props.element.type) {
-      case 'code':
-        return <CodeElement {...props} />;
-      default:
-        return <DefaultElement {...props} />;
-    }
-  }, []);
-
-  // Define a leaf rendering function that is memoized with `useCallback`.
-  const renderLeaf = useCallback((props) => <Leaf {...props} />, []);
-
-  return (
-    <Slate
-      editor={editor}
-      value={value}
-      onChange={(newValue) => {
-        setValue(newValue);
-        // Save the value to Local Storage.
-        const content = JSON.stringify(value);
-        localStorage.setItem('content', content);
-      }}
-    >
-      <div>
-        <button
-          onMouseDown={(event) => {
-            event.preventDefault();
-            CustomEditor.toggleBoldMark(editor);
-          }}
-        >
-          Bold
-        </button>
-        <button
-          onMouseDown={(event) => {
-            event.preventDefault();
-            CustomEditor.toggleCodeBlock(editor);
-          }}
-        >
-          Code Block
-        </button>
-      </div>
-      <Editable
-        renderElement={renderElement}
-        renderLeaf={renderLeaf}
-        onKeyDown={(event) => {
-          if (!event.ctrlKey) {
-            return;
-          }
-
-          switch (event.key) {
-            case '`': {
-              event.preventDefault();
-              CustomEditor.toggleCodeBlock(editor);
-              break;
-            }
-
-            case 'b': {
-              event.preventDefault();
-              CustomEditor.toggleBoldMark(editor);
-              break;
-            }
-          }
-        }}
-      />
-    </Slate>
-  );
-};
-
-export default TextEditor;
